update untuk fix bug

+++ resources/views/game/index.blade.php
@@ -49,10 +49,7 @@
                     <td>{{ $item->name }}</td>
                     <td>
                         <a href="/game/{{ $item->id }}/edit"> edit</a>
-                        <form id="deleteForm">
-                            <input type="hidden" id="id" name="id" required value="{{ $item->id }}"><br><br>
-                            <button type="submit">Delete</button>
-                        </form>
+                        <a data-id="{{ $item->id }}" class="btn btn-danger btn-hapus">hapus</a>
                     </td>
                 </tr>
             @endforeach
@@ -80,30 +77,26 @@
             });
         });
 
-        $(document).ready(function() {
-            $('#deleteForm').on('submit', function(e) {
-                e.preventDefault();
+        $(document).on('click', '.btn-hapus', function() {
+            const id = $(this).data('id')
+            const token = localStorage.getItem('token')
 
-                const id = $('#id').val();
+            confirm_dialog = confirm('Apakah anda yakin?');
 
+            if (confirm_dialog) {
                 $.ajax({
-                    url: `/api/game/${id}`,
-                    type: 'DELETE',
-                    data: {
-                        _token: '{{ csrf_token() }}' // Laravel CSRF token
+                    url: '/api/game/' + id,
+                    type: "DELETE",
+                    headers: {
+                        "Authorization": 'Bearer ' + token
                     },
                     success: function(response) {
-                        alert('Item deleted successfully!');
+                        alert('Form submitted successfully!');
                         console.log(response);
                     },
-                    error: function(xhr) {
-                        alert('An error occurred!');
-                        console.log(xhr.responseText);
-                    }
                 });
-            });
+            }
         });
-    </script>
     </script>
 </body>
